+++ packages/orq-rc/src/models/operations/createcontact.ts
@@ -134,7 +134,7 @@ export const CreateContactResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-02-27T11:09:03.160Z",
+    "2026-03-01T14:13:53.072Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {